Initializing a new project
type: chore

Starting a new project to pull and clean US wage data from the Bureau
of Labor Statistics. The current site and process to pull data is not
intuitive and the data sets pulled are incomplete and messy. Without
working in the sets often it is hard to get the information you need.

-Adding a README.md file