Build 1 of the Project
type: feat, chores

Adding in the bulk of the project. Started with the documentation and
labeling of each block of code for others to pick up easily and give
context to the process. Temporaarily reduced the span of data from 12
years to only 1 as we continue to test functionality and add in new
features. Below are the components added:

- Packages and File Path
- Pulling area definitions from BLS. This set of data helps a user to
merge wage statistics to areas and MSA which is how business usually
render this information
- Downloading the wage statistic zip files and extracting contents, then
building out dataframe with cleaned data based on the parameters
set by the user. The manageable dataframes allow users to continue
their analysis or export and distribute as is