Fixed Quarterly Bug
type: fix

There was an issue with the file paths in the code that was only pulling
folders ending in '4'. Was a a part that was included in the development
testing phase as all years' data was pulled through Q4. Now the code
can support any quarter's data when run.