Added Export to Excel
type: feat

After the report is clean results are now published to an Excel file in
the file path defined by the user at the start. This allows for data to
be access offline and filter by the paramerters necessary for one's
project. If needed again, users can open and read in the file using
Panda' read_excel() function which goes into a dataframe